requirements.txt
python-dotenv
requests
httpx
h2
